# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "jobinhood.co.uk API",
        "description": "",
        "version": ""
    },
    "paths": {
        "/api/user/login": {
            "post": {
                "summary": "Login the user",
                "operationId": "App\\Http\\Controllers\\Auth\\LoginController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@doe.com",
                                    "password": "Password",
                                    "device_name": "DeviceX"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token Issued"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "summary": "Adds a new user",
                "operationId": "App\\Http\\Controllers\\Auth\\RegisterController::register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@doe.com",
                                    "name": "John Doe",
                                    "password": "Password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    }
}